Financial statements formulas and welcome & country selection screens
Continuation of completing formulas in the financial statements for each model; pending verification of formulas at the end of the file and resolution of some formula-related doubts.

Welcome screen as the initial landing page for the tool.

Subsequent country selection screen to choose the implementation country.

Currently, only the Rwanda endpoint is included, as the tool development is based on this country.